requirements.txt
Flask==3.1.3
httpx==0.28.1
python-dotenv==1.2.1
gunicorn==23.0.0